fix(auth): scope /pod-token resolution by silo too (WeOwnAI no-workspace gate)
The WeOwnAI frontend establishes its OpenClaw session via POST
/auth/pod-token, not /auth/me. That broker (and /pod-token/cut) did the
same global, host-blind email->tenant lookup and fail-closed with 409
AMBIGUOUS_TENANT on >1 match — which the frontend renders as 'No
workspace for this account ... maps to more than one'. So a multi-silo
owner was blocked there too, independently of the /auth/me fix.

Scope both pod-token lookups to the silo derived from the request host,
mirroring /auth/me and the scope guard. Extract the host helpers
(_RequestHost, _ClusterTenantFromHost) into a shared infra/auth/request-silo
module so all four resolution paths use one implementation; oidc.service
now imports them instead of carrying private copies.

Tests: pod-token host-scoping case added; full control-plane suite green
(434).
