+++ apps/control-plane/src/infra/auth/oidc.service.ts
@@ -8,6 +8,7 @@ import type { PrismaClient } from "@prisma/client";
 
 import { ___LoadOidcAuthConfig } from "./oidc.config.js";
 import { _ResolveCallerClusterTenant } from "./resolve-caller-cluster-tenant.js";
+import { _ClusterTenantFromHost, _RequestHost } from "./request-silo.js";
 import { _ResolveOrgMembershipFacts } from "./org-membership.js";
 import type { OwnedOrg } from "./org-membership.js";
 
@@ -187,7 +188,7 @@ export class OidcAuthService
       //    OR-s the login-time flag (groups/operator) with membership-derived authority,
       //    so a user who just created an org is an org admin without re-logging-in.
       const [clusterTenant, membership] = await Promise.all([
-        this._resolveClusterTenant(authUser.email, _requestHost(req)),
+        this._resolveClusterTenant(authUser.email, _RequestHost(req)),
         _ResolveOrgMembershipFacts(this.prisma, authUser.sub),
       ]);
       return {
@@ -234,7 +235,7 @@ export class OidcAuthService
    */
   private async _resolveClusterTenant(email: string | undefined, host: string | undefined): Promise<string | null>
   {
-    return _ResolveCallerClusterTenant(this.prisma, email, _clusterTenantFromHost(host));
+    return _ResolveCallerClusterTenant(this.prisma, email, _ClusterTenantFromHost(host));
   }
 
   /** Build the provider redirect URL and persist PKCE state in the local session. */
@@ -576,34 +577,6 @@ export function ___CreateOidcAuthService(log: Logger, prisma: PrismaClient): Oid
 }
 
 /**
- * The request's effective host, honouring the `x-forwarded-host` set by the ingress proxy
- * (first value when comma-joined) and falling back to the `Host` header. Undefined when the
- * request carries no host. Shared by every per-org-host helper so they read the host one way.
- */
-function _requestHost(req: Request): string | undefined
-{
-  const forwardedHost = req.headers?.["x-forwarded-host"];
-  if (typeof forwardedHost === "string") return forwardedHost.split(",")[0].trim();
-  return typeof req.get === "function" ? req.get("host") : undefined;
-}
-
-/**
- * Derive the ClusterTenant (silo) the caller is currently on from the request host. Each org is
- * served at `<clusterTenant>.<base>`, so the first DNS label names the silo (port and case are
- * stripped). Returns undefined for a bare host with no subdomain (e.g. localhost) so the caller
- * falls back to an unscoped lookup. The label is only a candidate: the email→tenant query filters
- * on it and yields zero rows for a host whose first label is not a real silo, so a wrong guess
- * fail-closes rather than mis-resolving. Custom org domains that do not follow `<org>.<base>` are
- * not matched here (a future ingressHost-based lookup would cover them).
- */
-function _clusterTenantFromHost(host: string | undefined): string | undefined
-{
-  if (!host) return undefined;
-  const firstLabel = host.split(":")[0].trim().toLowerCase().split(".")[0];
-  return firstLabel || undefined;
-}
-
-/**
  * Build the OIDC redirect_uri for THIS request's host (DOMAIN.T4 multi-host). Each org is
  * served at its own host `<org>.<base>`, so login/callback must happen there for the
  * session cookie to be host-scoped to it. We take the callback PATH from the configured
@@ -616,7 +589,7 @@ function _buildRedirectUri(req: Request, configuredRedirect: string): string
 {
   const forwardedProto = req.headers["x-forwarded-proto"];
   const protocol = typeof forwardedProto === "string" ? forwardedProto.split(",")[0].trim() : req.protocol;
-  const host = _requestHost(req);
+  const host = _RequestHost(req);
   if (!host) return configuredRedirect;
   const callbackPath = new URL(configuredRedirect).pathname;
   return `${protocol}://${host}${callbackPath}`;
@@ -633,7 +606,7 @@ function _buildPostLogoutRedirectUri(req: Request, configuredRedirect: string): 
 {
   const forwardedProto = req.headers["x-forwarded-proto"];
   const protocol = typeof forwardedProto === "string" ? forwardedProto.split(",")[0].trim() : req.protocol;
-  const host = _requestHost(req);
+  const host = _RequestHost(req);
   if (!host) return configuredRedirect;
   const parsed = new URL(configuredRedirect);
   return `${protocol}://${host}${parsed.pathname}${parsed.search}`;
@@ -644,7 +617,7 @@ function _buildCurrentUrl(req: Request): URL
 {
   const forwardedProto = req.headers["x-forwarded-proto"];
   const protocol = typeof forwardedProto === "string" ? forwardedProto.split(",")[0].trim() : req.protocol;
-  const host = _requestHost(req);
+  const host = _RequestHost(req);
 
   return new URL(`${protocol}://${host}${req.originalUrl}`);
 }
